requirements.txt
gradio
langchain
langsmith
python-decouple
arxiv
pypdf2==3.0.1
tiktoken
langchain_openai
google-search-results
copyleaks
playwright 
beautifulsoup4
crewai
flask
supabase